Style NavigationLink with vanilla-extract
