+++ apps/store/src/components/Header/Header.css.ts
@@ -1,0 +1,10 @@
+import { style } from '@vanilla-extract/css'
+import { theme } from 'ui'
+
+export const focusableStyles = style({
+  cursor: 'pointer',
+
+  ':focus-visible': {
+    outline: `2px solid ${theme.colors.gray900}`,
+  },
+})
