+++ packages/ui/src/theme/radius.ts
@@ -1,10 +1,10 @@
 const base = {
-  0: 8,
-  1: 10,
-  2: 12,
-  3: 16,
-  4: 20,
-  5: 24,
+  0: '8px',
+  1: '10px',
+  2: '12px',
+  3: '16px',
+  4: '20px',
+  5: '24px',
 } as const
 
 export const radius = {
